refactor(blackjack): Split blackjack program into variants and views
Variants are stored in `variants/`
Views are stored in `views/`
