+++ casino/games/blackjack/__init__.py
@@ -1,3 +1,3 @@
-from .blackjack import play_blackjack
+from .variants import play_blackjack
 from .european import play_european_blackjack
 __all__ = ["play_blackjack", "play_european_blackjack"]
