+++ casino/games/blackjack/variants/__init__.py
@@ -1,0 +1,3 @@
+from .standard import StandardBlackjack, play_blackjack
+
+__all__ = ["StandardBlackjack", "play_blackjack"]
