+++ casino/games/blackjack/views/__init__.py
@@ -1,0 +1,3 @@
+from .view import BlackjackView
+
+__all__ = ["BlackjackView"]
